provide a parameter-less constructor for new users
Sadly, MoveItVisualTools(std::string) is already defined and
requires a *base_link*. In my opinion rather unfortunate, because the
topic seems much more likely to change, but that's what we got.
Instead, I opted for a constructor with hard-coded topic name.
Experienced users that need to change the topic should already know
the name of their base link.